build: node verision bump and security fixes
On the past PR the vercel bot shows that the deploy failed due to the end of the support of the 18.x
node versions, and the suggestion is to bump to the version 22.x. Since that was a necessary move,
it was also fixed some security issues moving the total amount of vulns to 5 low now.

BREAKING CHANGE: Since this is node version upddate and some security fixes, some side-effects might
happen